:recycle: Refactor: Repository 의존성 변경
<footer>
- 관련: #5
